Top picture updated dynamically
While playing episodes the top picture dynamically gets updated once the next episode starts playing
