+++ src/components/item-bar-episode.js
@@ -6,45 +6,25 @@ import useMediaPlayer from '../hooks/useMediaPlayer'
 const calcPercent = (dur) => (dur.position * 100 / dur.duration)
 
 const ItemBarEpisode = ({serie, episode, title, descr, useIcon, bkgrd, onClick}) => {
-  const { isPaused, curPlay, curPos } = useMediaPlayer()
+  const { curPos } = useMediaPlayer()
   const [mSec, setMSec] = useState(undefined)
   const [mSecDur, setMSecDur] = useState(undefined)
-  const restorePos = async (obj) => {
-    await apiObjGetStorage(obj,"mSec").then((value) => {
-      if (value==null){
-        value=0
-      }
-      setMSec(value)
-    }).catch((err) => {
-      console.error(err)
-    })
-    apiObjGetStorage(obj,"mSecDur").then((dur) => {
-      setMSecDur(dur)
-    }).catch((err) => {
-      console.error(err)
-    })
-  }
   useEffect(() => {
+    let didCancel = false
+    const restorePos = async (obj) => {
+      await apiObjGetStorage(obj,"mSec").then((value) => {
+        if (value==null)value=0
+        if (!didCancel) setMSec(value)
+      }).catch((err) => console.error(err))
+      await apiObjGetStorage(obj,"mSecDur").then((dur) => {
+        if (!didCancel) setMSecDur(dur)
+      }).catch((err) => console.error(err))
+    }
     if ((serie!=null) && (episode!=null)){
       restorePos({curSerie: serie, curEp: episode})
     }
+    return () => didCancel = true
   },[serie,episode])
-
-/*
-  componentWillReceiveProps = (nextProps) => {
-    const {episode,serie,curPos} = nextProps
-    if ((episode!=null) && (serie!=null)){
-      if (episode!==this.props.episode){
-        this.restoreCurPos({curSerie: serie, curEp: episode})
-      } else if (this.props.isActive
-                  && (curPos!=null)
-                  && (curPos.position!==this.props.position)
-                  && (curPos.position!==this.state.mSec)){
-        this.setState({mSec: curPos.position, mSecDur: curPos.duration})
-      }
-    }
-  }
-*/
   const handleClick = (ev, idStr) => {
     const resetPosMargin = 10000 // Reset playing to begining if less mSec remains
     ev.stopPropagation()
